+++ compiler/qsc_runtime_capabilities/src/analysis_legacy.rs
